requirements.txt
asgiref==3.8.1
Django==4.1.11
djangorestframework==3.15.2
mysqlclient==2.2.7
pillow==11.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
